third_party/WebKit/LayoutTests/http/tests/csspaint/dynamic-parse-insecure-context.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/get-host-info.js"></script>
<style>
#test {
  background: green;
}
</style>
<div id="test"></div>
<script>
  if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
    window.location = get_host_info().UNAUTHENTICATED_ORIGIN + window.location.pathname;
  } else {
    test(function() {
      const GREEN = 'rgb(0, 128, 0)';
      const element = document.getElementById('test');
      assert_false(window.isSecureContext);

      element.setAttribute('style', 'background: paint(fail) purple');
      assert_equals(getComputedStyle(element).backgroundColor, GREEN);

      element.style.background = 'paint(fail) purple';
      assert_equals(getComputedStyle(element).backgroundColor, GREEN);

      document.styleSheets[0].insertRule(
          '#test { background: paint(fail) purple; }');
      assert_equals(getComputedStyle(element).backgroundColor, GREEN);

      document.styleSheets[0].insertRule(
          '#test { --bg: paint(fail) purple; background: var(--bg); }');
      assert_equals(getComputedStyle(element).backgroundColor, GREEN);
    });
  }
</script>
